refactor(shell): extract shared CompanionSwitcher; use in mobile nav
